+++ slotlist.h
@@ -17,11 +17,11 @@ public:
     int slotInterval;
     int labelCount;
     int currentIndex = 0;
-    long currentTime = 0;
+    long long currentTime = 0;
 
     SlotList(int listId, int slotCount, int slotInterval, int labelCount);
 
-    long getMaxBytes();
+    long long getMaxBytes();
     Slot *get(int index);
     void setCurrentTime(long long currentTime);
     void update(int slotId, long long startTime, long long rxBytes, long long txBytes);
